[Refactoring] 파일 이름 규칙 적용
