+++ EmployeeManagementSystem/EMSGTest/MockManage.h
@@ -2,7 +2,7 @@
 
 #include "pch.h"
 #include <stdexcept>
-#include "../EmployeeManagementSystem/employee.h"
+#include "../EmployeeManagementSystem/Employee.h"
 #include "../EmployeeManagementSystem/EmployeeManagementSystem.h"
 using namespace std;
